backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pandas
scikit-learn
shap>=0.41.0
python-multipart
alembic
passlib[bcrypt]
PyJWT
pdfplumber
reportlab
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
mlflow==2.22.0
pytest
pytest-cov
protobuf>=4.25.0
